added user id checking

+++ src/main/java/ru/chausov/to_do_list/servlet/TaskController.java
@@ -4,7 +4,9 @@ package ru.chausov.to_do_list.servlet;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 import ru.chausov.to_do_list.data_base.entity.Task;
 import ru.chausov.to_do_list.data_base.entity.User;
@@ -14,6 +16,7 @@ import ru.chausov.to_do_list.data_base.repository.UserRepository;
 import javax.transaction.Transactional;
 import java.security.Principal;
 import java.time.LocalDate;
+import java.util.List;
 import java.util.Map;
 
 
@@ -54,38 +57,76 @@ public class TaskController {
 
     @Transactional
     @PostMapping("/delete")
-    public ModelAndView deleteTask(Long id) {
-        taskRepository.deleteById(id);
+    public ModelAndView deleteTask(Principal authUser, Long id) {
+        User user = userRepository.findByUsername(authUser.getName());
+
+        List<Task> tasks = user.getTasks();
+
+        boolean found = false;
+
+        for(Task taskToCheck : tasks) {
+            if(taskToCheck.getId().equals(id))
+                found = true;
+        }
+
+        if(found)
+            taskRepository.deleteById(id);
 
         return new ModelAndView("redirect:/tasks/table");
     }
 
     @Transactional
     @PostMapping("/update")
-    public ModelAndView updateTask(Long id, Task task, Map<String, Object> model) {
-        Task taskToUpdate = taskRepository.findById(id).get();
+    public ModelAndView updateTask(Principal authUser, Long id, Task task, Map<String, Object> model) {
+        User user = userRepository.findByUsername(authUser.getName());
 
-        taskToUpdate.setName(task.getName());
-        taskToUpdate.setDescription(task.getDescription());
-        taskToUpdate.setReceivedDate(task.getReceivedDate());
-        taskToUpdate.setToBeDone(task.getToBeDone());
+        List<Task> tasks = user.getTasks();
 
-        model.put("task", taskRepository.save(taskToUpdate));
+        boolean found = false;
+
+        for(Task taskToCheck : tasks) {
+            if(taskToCheck.getId().equals(id))
+                found = true;
+        }
+
+        if(found) {
+            Task taskToUpdate = taskRepository.findById(id).get();
+
+            taskToUpdate.setName(task.getName());
+            taskToUpdate.setDescription(task.getDescription());
+            taskToUpdate.setReceivedDate(task.getReceivedDate());
+            taskToUpdate.setToBeDone(task.getToBeDone());
+
+            model.put("task", taskRepository.save(taskToUpdate));
+        }
 
         return new ModelAndView("redirect:/tasks/table", model);
     }
 
     @Transactional
     @PostMapping("/done")
-    public ModelAndView setTaskDone(Long id, Map<String, Object> model) {
-        Task task = taskRepository.findById(id).get();
+    public ModelAndView setTaskDone(Principal authUser, Long id, Map<String, Object> model) {
+        User user = userRepository.findByUsername(authUser.getName());
 
-        if(!task.isDone())
-            task.setDone(true);
-        else
-            task.setDone(false);
+        List<Task> tasks = user.getTasks();
 
-        model.put("task", taskRepository.save(task));
+        boolean found = false;
+
+        for(Task taskToCheck : tasks) {
+            if(taskToCheck.getId().equals(id))
+                found = true;
+        }
+
+        if(found) {
+            Task task = taskRepository.findById(id).get();
+
+            if (!task.isDone())
+                task.setDone(true);
+            else
+                task.setDone(false);
+
+            model.put("task", taskRepository.save(task));
+        }
 
         return new ModelAndView("redirect:/tasks/table", model);
     }
